Get rid of Show and Pretty instances for AST types
The only visualization that is necessary for users is of types, and
that is moving to Reporting.Render.Type where we’ll also have “type
diffs” that’ll make it easier to see how two different types do not fit
together.

It may be necessary to add some of this back later for debugging
purposes, but it needs to be rewritten either way. I’m just pushing
that down the line until I actually need to do it.
