+++ elm-compiler.cabal
@@ -108,7 +108,6 @@ Library
         Reporting.Error.Pattern,
         Reporting.Error.Syntax,
         Reporting.Error.Type,
-        Reporting.PrettyPrint,
         Reporting.Region,
         Reporting.Report,
         Reporting.Result,
@@ -130,6 +129,7 @@ Library
         aeson >= 0.7 && < 0.9,
         aeson-pretty >= 0.7 && < 0.8,
         ansi-terminal >= 0.6.2.1 && < 0.7,
+        ansi-wl-pprint >= 0.6.7 && < 0.7,
         base >=4.2 && <5,
         binary >= 0.7.0.0 && < 0.8,
         bytestring >= 0.9 && < 0.11,
@@ -189,7 +189,6 @@ Test-Suite compiler-tests
         Test.Property.Arbitrary
         AST.Helpers
         AST.Literal
-        Reporting.PrettyPrint
 
     build-depends:
         test-framework > 0.8 && < 0.9,
@@ -200,6 +199,7 @@ Test-Suite compiler-tests
         aeson >= 0.7 && < 0.9,
         aeson-pretty >= 0.7 && < 0.8,
         ansi-terminal >= 0.6.2.1 && < 0.7,
+        ansi-wl-pprint >= 0.6.7 && < 0.7,
         base >=4.2 && <5,
         binary >= 0.7.0.0 && < 0.8,
         bytestring >= 0.9 && < 0.11,
