+++ src/AST/Pattern.hs
@@ -2,13 +2,10 @@
 module AST.Pattern where
 
 import qualified Data.Set as Set
-import Text.PrettyPrint as P
 
-import qualified AST.Helpers as Help
 import qualified AST.Literal as L
 import qualified AST.Variable as Var
 import qualified Reporting.Annotation as A
-import qualified Reporting.PrettyPrint as P
 import qualified Reporting.Region as R
 
 
@@ -23,7 +20,6 @@ data Pattern' ann var
     | Var String
     | Anything
     | Literal L.Literal
-    deriving (Show)
 
 
 type RawPattern =
@@ -61,7 +57,9 @@ tuple patterns =
   Data (Var.Raw ("_Tuple" ++ show (length patterns))) patterns
 
 
+
 -- FIND VARIABLES
+
 
 boundVars :: Pattern ann var -> [A.Annotated ann String]
 boundVars (A.A ann pattern) =
@@ -99,45 +97,3 @@ boundVarList :: Pattern ann var -> [String]
 boundVarList pattern =
   Set.toList (boundVarSet pattern)
 
-
--- PRETTY PRINTING
-
-instance Var.ToString var => P.Pretty (Pattern' ann var) where
-  pretty dealiaser needsParens pattern =
-    case pattern of
-      Var name ->
-          P.text name
-
-      Literal literal ->
-          P.pretty dealiaser needsParens literal
-
-      Record fields ->
-          P.braces (P.commaCat (map P.text fields))
-
-      Alias x ptrn ->
-          P.parensIf needsParens $
-              P.pretty dealiaser True ptrn <+> P.text "as" <+> P.text x
-
-      Anything ->
-          P.text "_"
-
-      Data name [A.A _ hd, A.A _ tl]
-          | Var.toString name == "::" ->
-              P.parensIf isCons (P.pretty dealiaser False hd)
-              <+> P.text "::"
-              <+> P.pretty dealiaser False tl
-          where
-            isCons =
-              case hd of
-                Data ctor _ -> Var.toString ctor == "::"
-                _ -> False
-
-      Data name patterns ->
-          let name' = Var.toString name
-          in
-            if Help.isTuple name'
-              then
-                P.parens (P.commaCat (map (P.pretty dealiaser False) patterns))
-              else
-                P.parensIf needsParens $
-                    P.hsep (P.text name' : map (P.pretty dealiaser True) patterns)
