+++ src/Elm/Compiler.hs
@@ -6,7 +6,8 @@ module Elm.Compiler
     , Dealiaser, dummyDealiaser
     , Error, errorToString, errorToJson, printError
     , Warning, warningToString, warningToJson, printWarning
-    ) where
+    )
+    where
 
 import qualified Data.Aeson as Json
 import qualified Data.Map as Map
@@ -24,7 +25,7 @@ import qualified Parse.Module as Parse
 import qualified Parse.Parse as Parse
 import qualified Reporting.Annotation as A
 import qualified Reporting.Error as Error
-import qualified Reporting.PrettyPrint as P
+import qualified Reporting.Render.Type as RenderType
 import qualified Reporting.Result as Result
 import qualified Reporting.Warning as Warning
 
@@ -71,7 +72,7 @@ compile context source interfaces =
     (Context packageName isRoot isExposed dependencies) =
       context
 
-    (Result.Result (dealiaser, warnings) rawResult) =
+    (Result.Result (localizer, warnings) rawResult) =
       do  modul <- Compile.compile packageName isRoot dependencies interfaces source
           docs <- docsGen isExposed modul
 
@@ -80,7 +81,7 @@ compile context source interfaces =
 
           return (Result docs interface javascript)
   in
-    ( maybe dummyDealiaser Dealiaser dealiaser
+    ( maybe dummyDealiaser Dealiaser localizer
     , map Warning warnings
     , Result.destruct (Left . map Error) Right rawResult
     )
@@ -125,7 +126,7 @@ docsGen isExposed modul =
 -- DEALIASER
 
 newtype Dealiaser =
-    Dealiaser P.Dealiaser
+    Dealiaser RenderType.Localizer
 
 
 dummyDealiaser :: Dealiaser
@@ -140,18 +141,18 @@ newtype Error =
 
 
 errorToString :: Dealiaser -> String -> String -> Error -> String
-errorToString (Dealiaser dealiaser) location source (Error err) =
-    Error.toString dealiaser location source err
+errorToString (Dealiaser localizer) location source (Error err) =
+    Error.toString localizer location source err
 
 
 printError :: Dealiaser -> String -> String -> Error -> IO ()
-printError (Dealiaser dealiaser) location source (Error err) =
-    Error.print dealiaser location source err
+printError (Dealiaser localizer) location source (Error err) =
+    Error.print localizer location source err
 
 
 errorToJson :: Dealiaser -> String -> Error -> Json.Value
-errorToJson (Dealiaser dealiaser) location (Error err) =
-    Error.toJson dealiaser location err
+errorToJson (Dealiaser localizer) location (Error err) =
+    Error.toJson localizer location err
 
 
 -- WARNINGS
@@ -161,17 +162,17 @@ newtype Warning =
 
 
 warningToString :: Dealiaser -> String -> String -> Warning -> String
-warningToString (Dealiaser dealiaser) location source (Warning err) =
-    Warning.toString dealiaser location source err
+warningToString (Dealiaser localizer) location source (Warning err) =
+    Warning.toString localizer location source err
 
 
 printWarning :: Dealiaser -> String -> String -> Warning -> IO ()
-printWarning (Dealiaser dealiaser) location source (Warning err) =
-    Warning.print dealiaser location source err
+printWarning (Dealiaser localizer) location source (Warning err) =
+    Warning.print localizer location source err
 
 
 warningToJson :: Dealiaser -> String -> Warning -> Json.Value
-warningToJson (Dealiaser dealiaser) location (Warning err) =
-    Warning.toJson dealiaser location err
+warningToJson (Dealiaser localizer) location (Warning err) =
+    Warning.toJson localizer location err
