+++ src/Reporting/Warning.hs
@@ -4,15 +4,12 @@ module Reporting.Warning where
 
 import Data.Aeson ((.=))
 import qualified Data.Aeson as Json
-import qualified Text.PrettyPrint as P
-import Text.PrettyPrint ((<+>))
 
-import qualified AST.Helpers as Help
 import qualified AST.Module.Name as ModuleName
 import qualified AST.Type as Type
 import qualified Reporting.Annotation as A
-import qualified Reporting.PrettyPrint as P
 import qualified Reporting.Report as Report
+import qualified Reporting.Render.Type as RenderType
 
 
 -- ALL POSSIBLE WARNINGS
@@ -24,18 +21,18 @@ data Warning
 
 -- TO STRING
 
-toString :: P.Dealiaser -> String -> String -> A.Located Warning -> String
-toString dealiaser location source (A.A region warning) =
-    Report.toString location region (toReport dealiaser warning) source
+toString :: RenderType.Localizer -> String -> String -> A.Located Warning -> String
+toString localizer location source (A.A region warning) =
+    Report.toString location region (toReport localizer warning) source
 
 
-print :: P.Dealiaser -> String -> String -> A.Located Warning -> IO ()
-print dealiaser location source (A.A region warning) =
-    Report.printWarning location region (toReport dealiaser warning) source
+print :: RenderType.Localizer -> String -> String -> A.Located Warning -> IO ()
+print localizer location source (A.A region warning) =
+    Report.printWarning location region (toReport localizer warning) source
 
 
-toReport :: P.Dealiaser -> Warning -> Report.Report
-toReport dealiaser warning =
+toReport :: RenderType.Localizer -> Warning -> Report.Report
+toReport localizer warning =
   case warning of
     UnusedImport moduleName ->
         Report.simple
@@ -48,29 +45,18 @@ toReport dealiaser warning =
           "missing type annotation"
           ("Top-level value `" ++ name ++ "` does not have a type annotation.")
           ( "The type annotation you want looks something like this:\n\n"
-            ++ P.render (P.nest 4 typeDoc)
+            ++ RenderType.annotation name inferredType
           )
-      where
-        nameDoc =
-          if Help.isOp name then
-              P.parens (P.text name)
-          else
-              P.text name
 
-        typeDoc =
-          P.hang
-            (nameDoc <+> P.colon)
-            4
-            (P.pretty dealiaser False inferredType)
 
 
 -- TO JSON
 
-toJson :: P.Dealiaser -> FilePath -> A.Located Warning -> Json.Value
-toJson dealiaser filePath (A.A region warning) =
+toJson :: RenderType.Localizer -> FilePath -> A.Located Warning -> Json.Value
+toJson localizer filePath (A.A region warning) =
   let
     (maybeRegion, additionalFields) =
-        Report.toJson [] (toReport dealiaser warning)
+        Report.toJson [] (toReport localizer warning)
   in
       Json.object $
         [ "file" .= filePath
